Move content-api utilities and instances creation into their own file, add jsdoc
